+++ tests/issues/issue_937.py
@@ -1,10 +1,6 @@
 ############################ Copyrights and license ############################
 #                                                                              #
-# Copyright 2013 Vincent Jacques <vincent@vincent-jacques.net>                 #
-# Copyright 2014 Vincent Jacques <vincent@vincent-jacques.net>                 #
-# Copyright 2016 Peter Buckley <dx-pbuckley@users.noreply.github.com>          #
-# Copyright 2018 Steve Kowalik <steven@wedontsleep.org>                        #
-# Copyright 2018 sfdye <tsfdye@gmail.com>                                      #
+# Copyright 2018 Vinay Hegde <vinayhegde2010@gmail.com>                        #
 #                                                                              #
 # This file is part of PyGithub.                                               #
 # http://pygithub.readthedocs.io/                                              #
@@ -23,31 +19,17 @@
 # along with PyGithub. If not, see <http://www.gnu.org/licenses/>.             #
 #                                                                              #
 ################################################################################
-
-import github
-
-from . import Framework
+from tests import Framework
 
 
-class Issue134(
-    Framework.BasicTestCase
-):  # https://github.com/jacquev6/PyGithub/pull/134
-    def testGetAuthorizationsFailsWhenAutenticatedThroughOAuth(self):
-        g = github.Github(auth=self.oauth_token)
-        with self.assertRaises(github.GithubException) as raisedexp:
-            list(g.get_user().get_authorizations())
-        self.assertEqual(raisedexp.exception.status, 404)
+class Issue937(Framework.TestCase):
+    def setUp(self):
+        super().setUp()
+        self.user = self.g.get_user()
+        self.repo = self.user.get_repo("PyGithub")
 
-    def testGetAuthorizationsSucceedsWhenAutenticatedThroughLoginPassword(self):
-        g = github.Github(auth=self.login)
-        self.assertListKeyEqual(
-            g.get_user().get_authorizations(),
-            lambda a: a.note,
-            [None, None, "cligh", None, None, "GitHub Android App"],
-        )
-
-    def testGetOAuthScopesFromHeader(self):
-        g = github.Github(auth=self.oauth_token)
-        self.assertEqual(g.oauth_scopes, None)
-        g.get_user().name
-        self.assertEqual(g.oauth_scopes, ["repo", "user", "gist"])
+    def testCollaboratorsAffiliation(self):
+        collaborators = self.repo.get_collaborators(affiliation="direct")
+        self.assertListKeyEqual(collaborators, lambda u: u.login, ["hegde5"])
+        with self.assertRaises(AssertionError):
+            self.repo.get_collaborators(affiliation="invalid_option")
